refactor(adapters): extract shared kysely/drizzle helpers into adapter-interfaces
The date-conversion helpers, SQL row/entity transforms, and the Lucene
query sanitizer were maintained as verbatim copies in both SQL adapter
packages (and had already started to drift in comments). Move the
implementations to @authhero/adapter-interfaces — the canonical home for
cross-adapter code, alongside base64url and the cursor codec — and turn
the kysely/drizzle modules into re-exports so all existing import paths
keep working.
