+++ packages/adapter-interfaces/src/utils/index.ts
@@ -4,3 +4,6 @@ export * from "./connection-attributes";
 export * from "./guards";
 export * from "./base64url";
 export * from "./cursor";
+export * from "./dates";
+export * from "./sql-transform";
+export * from "./lucene";
