+++ packages/kysely/src/helpers/filter.ts
@@ -27,62 +27,9 @@ function unescapeLuceneValue(value: string): string {
   return value.replace(/\\([\\"+\-!(){}[\]^~*?:/&|])/g, "$1");
 }
 
-// Strip field-scoped clauses (`field:value`, `-field:value`, `_exists_:field`,
-// `field=value`) whose field is not in `allowedFields`. Bare-string tokens are
-// preserved (luceneFilter routes them through its own searchable-columns
-// whitelist). Returns a query string safe to pass into luceneFilter.
-export function sanitizeLuceneQuery(
-  query: string,
-  allowedFields: string[],
-): string {
-  const allowed = new Set(allowedFields);
-
-  const sanitizePart = (part: string): string => {
-    const tokens: string[] = [];
-    let current = "";
-    let inQuotes = false;
-    for (let i = 0; i < part.length; i++) {
-      const char = part[i];
-      if (char === '"') {
-        inQuotes = !inQuotes;
-        current += char;
-      } else if (char === " " && !inQuotes) {
-        if (current.trim()) tokens.push(current.trim());
-        current = "";
-      } else {
-        current += char;
-      }
-    }
-    if (current.trim()) tokens.push(current.trim());
-
-    const kept = tokens.filter((token) => {
-      const normalized = token.replace(/^([^:]+)=/, "$1:");
-      const stripped = normalized.startsWith("-")
-        ? normalized.slice(1)
-        : normalized;
-
-      if (stripped.startsWith("_exists_:")) {
-        return allowed.has(stripped.slice(9));
-      }
-      const colonIdx = stripped.indexOf(":");
-      if (colonIdx > 0) {
-        return allowed.has(stripped.slice(0, colonIdx));
-      }
-      return true;
-    });
-
-    return kept.join(" ");
-  };
-
-  const orParts = query.split(/ OR /i);
-  if (orParts.length > 1) {
-    return orParts
-      .map(sanitizePart)
-      .filter((p) => p.length > 0)
-      .join(" OR ");
-  }
-  return sanitizePart(query);
-}
+// Re-exported so existing `../helpers/filter` imports keep working; the
+// implementation is shared with the drizzle adapter via adapter-interfaces.
+export { sanitizeLuceneQuery } from "@authhero/adapter-interfaces";
 
 // Generic over the query builder's DB type (not just `Database`) because
 // left-joined builders carry a widened DB type with nullable joined columns.
@@ -143,11 +90,7 @@ export function luceneFilter<DB, TB extends keyof DB, O>(
             if (likeSet.has(fieldName)) {
               return eb(ref(toColumn(fieldName)), "like", `%${cleanValue}%`);
             }
-            return eb(
-              toLhs(fieldName),
-              "=",
-              toOperand(fieldName, cleanValue),
-            );
+            return eb(toLhs(fieldName), "=", toOperand(fieldName, cleanValue));
           }
           return null;
         })
