+++ packages/kysely/src/utils/getCountAsInt.ts
@@ -1,12 +1,4 @@
-export default function getCountAsInt(count: string | number | bigint) {
-  // VScode complains that parseInt only accepts a string BUT the project builds & lints
-  if (typeof count === "string") {
-    return parseInt(count, 10);
-  }
-
-  if (typeof count === "bigint") {
-    return Number(count);
-  }
-
-  return count;
-}
+// Re-exported from @authhero/adapter-interfaces, the canonical home for
+// helpers shared between the SQL adapters. Kept as a default export because
+// all existing imports use `import getCountAsInt from ...`.
+export { getCountAsInt as default } from "@authhero/adapter-interfaces";
